Handle objects in stringifyJSON
Handle empty arrays as base case.  Add recursive
case that iterates through the keys and creates a
string composed of the string version of the key,
a : and the string version of the key-value. Store
these key value strings in an array and then return
the array with a { and }.  If the stringified key or
value is undefined or a function, remove the key
value string.