Bulk of getElementsByClassName
This work gets close to returning the expected
value I believe.  It creates an array of the child
nodes and then iterates through each with the last
index and pop, adding the node to a helper array
if it contains the desired class name.  A problem
still exists with matching the result to expected,
possibly due to order in array?