Pass all tests for getElementsByClassName
This work makes sure to also look for child nodes
within child nodes that might match the class name.
I am concerned though that this solution is
inefficient since it creates a large array first
and then evaluates it.  It would probably be more
efficient to evaluate while creating the list...